Allow for Tranco Automation
This PR updates and renames the existing `process-tranco` workflow to run on a schedule (the first of the month) and create a PR with the required changes. 

The workflow also allows for manual triggering should it be required. 

You can see the action running within my fork: https://github.com/zoomequipd/static-files/actions/runs/14225806235/job/39864931769

Special handling of curl was required due to frequent HTTP 522 responses when fetching the ID of the current Tranco list.